Util: replace CharSet.for("UTF-8") with StandardCharsets.UTF_8 and suppress DM_DEFAULT_ENCODING for console input
Default system encoding should be kept for Scanner considering users might be using non-UTF8 passwords
