+++ src/main/java/org/semux/api/ConsoleApi.java
@@ -13,9 +13,10 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 
+import org.semux.api.response.GetBlockResponse;
+
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import org.semux.api.response.GetBlockResponse;
 
 /**
  * Additional console-only API
