+++ src/main/java/org/semux/util/SystemUtil.java
@@ -6,17 +6,16 @@
  */
 package org.semux.util;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
+
 import java.io.BufferedReader;
-import java.io.Console;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.net.URL;
 import java.net.URLConnection;
-import java.nio.charset.Charset;
 import java.util.Locale;
-import java.util.Scanner;
 import java.util.concurrent.ExecutionException;
 
 import org.semux.config.Constants;
@@ -47,8 +46,6 @@ public class SystemUtil {
     }
 
     private static final Logger logger = LoggerFactory.getLogger(SystemUtil.class);
-
-    private static final Scanner scanner = new Scanner(new InputStreamReader(System.in, Charset.forName("UTF-8")));
 
     public enum OsName {
         WINDOWS("Windows"),
@@ -136,7 +133,7 @@ public class SystemUtil {
             con.setReadTimeout(Constants.DEFAULT_READ_TIMEOUT);
 
             BufferedReader reader = new BufferedReader(
-                    new InputStreamReader(con.getInputStream(), Charset.forName("UTF-8")));
+                    new InputStreamReader(con.getInputStream(), UTF_8));
             String ip = reader.readLine().trim();
             reader.close();
 
@@ -155,52 +152,6 @@ public class SystemUtil {
         } catch (Exception e) {
             return InetAddress.getLoopbackAddress().getHostAddress();
         }
-    }
-
-    /**
-     * Reads a line from the console.
-     * 
-     * @param prompt
-     * @return
-     */
-    public static String readLine(String prompt) {
-        if (prompt != null) {
-            System.out.print(prompt);
-            System.out.flush();
-        }
-
-        return scanner.nextLine();
-    }
-
-    /**
-     * Reads a password from the console.
-     *
-     * @param prompt
-     *            A message to display before reading password
-     * @return
-     */
-    public static String readPassword(String prompt) {
-        Console console = System.console();
-
-        if (console == null) {
-            if (prompt != null) {
-                System.out.print(prompt);
-                System.out.flush();
-            }
-
-            return scanner.nextLine();
-        }
-
-        return new String(console.readPassword(prompt));
-    }
-
-    /**
-     * Reads a password from the console.
-     *
-     * @return
-     */
-    public static String readPassword() {
-        return readPassword("Please enter your password: ");
     }
 
     /**
