+++ src/test/java/org/semux/gui/SwingUtilTest.java
@@ -6,9 +6,9 @@
  */
 package org.semux.gui;
 
-import java.math.BigDecimal;
 import static org.junit.Assert.assertEquals;
 
+import java.math.BigDecimal;
 import java.text.ParseException;
 import java.util.Locale;
